requirements.txt
amqp==5.1.1
anyjson==0.3.3
argh==0.27.2
asgiref==3.4.1
asn1crypto==1.5.1
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configglue==1.2
#crispy-bootstrap4==2022.1
cron-descriptor==1.2.35
cryptography==40.0.1
Cython==0.26
decorator==5.1.1
#deliac==1.0.14
dj-database-url==2.0.0
Django==3.2.19
django-celery-beat==2.0.0
django-configglue==0.7.1
django-configurations==2.3.2
django-crispy-forms==1.13.0
django-dotenv==1.4.2
django-environ==0.10.0
django-filter==21.1
django-jsonfield==0.9.19
django-sendfile==0.3.10
django-timezone-field==4.2.3
djangorestframework==3.14.0
Events==0.4
gunicorn==19.4.5
idna==3.4
invoke==2.0.0
ipython==4.0.1
joblib==1.1.1
jsonfield==3.1.0
kombu==5.1.0
ldap3==2.2.1
medusa==0.5.4
meld3==1.0.2
nltk==3.6.7
paramiko==3.1.0
pathtools==0.1.2
pexpect==4.8.0
pg8000==1.26.0
pickleshare==0.7.5
polib==1.2.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
PyNaCl==1.5.0
pysqlite3==0.5.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
pyxdg==0.28
PyYAML==6.0
regex==2023.5.5
requests==2.27.1
ruby==0.0.2
scramp==1.4.1
simplegeneric==0.8.1
six==1.16.0
sqlparse==0.4.3
supervisor==4.2.5
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
virtualenv==14.0.6
watchdog==0.8.3
wcwidth==0.2.6
XlsxWriter==0.8.4
